feat(rust): TcpTransport::stop_outlet, examples
- Add `stop_outlet()` to TcpTransport and TcpRouterHandle
- Add examples to TcpTransport functions
- Fix clippy warnings
